pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "gptparse"
version = "0.1.0"
description = "A tool for converting PDF documents to Markdown using OCR and vision language models"
authors = ["Sanjay Nadhavajhala <sanjay@acorn.io>"]
readme = "README.md"
repository = "https://github.com/gptscript-ai/gptparse"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.9"
click = "8.1.7"
pdf2image = "1.17.0"
Pillow = "10.4.0"
PyPDF2 = "3.0.1"
requests = "2.32.3"
tqdm = "4.66.5"
langchain-anthropic = "0.2.3"
langchain-openai = "0.2.2"
langchain-google-genai = "2.0.1"
beautifulsoup4 = "4.12.3"
cffi = "1.17.1"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
decorator = "5.1.1"
filelock = "3.16.1"
fsspec = "2024.9.0"
idna = "3.10"
platformdirs = "4.3.6"
psutil = "6.0.0"
PyYAML = "6.0.2"
six = "1.16.0"
traitlets = "5.14.3"
google-ai-generativelanguage = "0.6.10"
google-api-core = "2.21.0"
google-auth = "2.35.0"
google-generativeai = "0.8.3"
huggingface-hub = "0.25.2"
openai = "1.51.2"
grpcio = "1.66.2"
grpcio-status = "1.66.2"
httpx = "0.27.2"
httpcore = "1.0.6"
markdown-it-py = "3.0.0"
mistune = "3.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[tool.poetry.scripts]
gptparse = "gptparse.cli:main"